Clang-Tidy V (#3201)
This is the fifth batch of changes fixing clang-tidy warnings. The
changes are most related to single precision, 1D, 2D, and OMP.

Enable clang-tidy in GCC and Clang CIs.

Switch to Ubuntu 22.04 and GCC 12 in one test.

+++ Src/Base/AMReX_BoxList.cpp
@@ -472,7 +472,7 @@ BoxList::parallelComplementIn (const Box& b, BoxArray const& ba)
                     ba.complementIn(bl_tmp, bl_mesh.m_lbox[i]);
                     vbox.insert(std::end(vbox), std::begin(bl_tmp), std::end(bl_tmp));
                 }
-                ntot += bl_tmp.size();
+                ntot += static_cast<int>(bl_tmp.size());
             }
             local_boxes.reserve(ntot);
             for (auto& bl : bl_priv) {
